server/requirements.txt
python-dotenv
pymysql
sqlalchemy
fastapi
strawberry-graphql[asgi]